доработка кода2
